Organize UGNI header files
