+++ src/uct/ugni/base/ugni_ep.c
@@ -3,8 +3,8 @@
  * Copyright (C) Mellanox Technologies Ltd. 2001-2014.  ALL RIGHTS RESERVED.
  * See file LICENSE for terms.
  */
-#include <uct/ugni/base/ugni_ep.h>
-#include <uct/ugni/base/ugni_iface.h>
+#include "ugni_ep.h"
+#include "ugni_iface.h"
 
 SGLIB_DEFINE_LIST_FUNCTIONS(uct_ugni_ep_t, uct_ugni_ep_compare, next);
 SGLIB_DEFINE_HASHED_CONTAINER_FUNCTIONS(uct_ugni_ep_t, UCT_UGNI_HASH_SIZE, uct_ugni_ep_hash);
