+++ src/uct/ugni/base/ugni_iface.c
@@ -1,3 +1,7 @@
+#include "ugni_types.h"
+#include "ugni_md.h"
+#include "ugni_device.h"
+#include "ugni_ep.h"
 #include "ugni_iface.h"
 #include <pmi.h>
 
@@ -287,8 +291,6 @@ ucs_status_t uct_ugni_init_nic(int device_index,
     ++ugni_domain_global_counter;
     return UCS_OK;
 }
-
-#define UCT_UGNI_LOCAL_CQ (8192)
 
 ucs_status_t ugni_activate_iface(uct_ugni_iface_t *iface)
 {
