+++ src/uct/ugni/smsg/ugni_smsg_iface.c
@@ -4,13 +4,11 @@
  * See file LICENSE for terms.
  */
 
-#include <pmi.h>
-#include "ucs/type/class.h"
-
-#include <ucs/arch/cpu.h>
-#include <uct/ugni/base/ugni_iface.h>
 #include "ugni_smsg_iface.h"
 #include "ugni_smsg_ep.h"
+#include <uct/ugni/base/ugni_md.h>
+#include <uct/ugni/base/ugni_device.h>
+#include <ucs/arch/cpu.h>
 
 #define UCT_UGNI_SMSG_TL_NAME "ugni_smsg"
