+++ src/uct/ugni/udt/ugni_udt_iface.h
@@ -4,16 +4,13 @@
  * See file LICENSE for terms.
  */
 
-#ifndef UCT_UGNI_IFACE_H
-#define UCT_UGNI_IFACE_H
+#ifndef UCT_UGNI_UDT_IFACE_H
+#define UCT_UGNI_UDT_IFACE_H
 
-#include <gni_pub.h>
-#include <uct/ugni/base/ugni_md.h>
-#include <uct/ugni/base/ugni_device.h>
+#include "ugni_udt_ep.h"
+#include <uct/ugni/base/ugni_types.h>
 #include <uct/ugni/base/ugni_iface.h>
 #include <ucs/datastruct/list.h>
-#include "ugni_udt_ep.h"
-
 #include <uct/base/uct_md.h>
 #include <ucs/async/async.h>
 #include <ucs/async/pipe.h>
@@ -51,6 +48,7 @@ typedef struct uct_ugni_udt_header {
 } uct_ugni_udt_header_t;
 
 void uct_ugni_udt_progress(void *arg);
+
 #define uct_ugni_udt_get_offset(i) ((size_t)(ucs_max(sizeof(uct_ugni_udt_header_t), ((i)->config.rx_headroom  + \
                  sizeof(uct_recv_desc_t)))))
